Brings the radiative-transfer (RT) / spatial-angular AMR work from
yt/Radiation_newmaster onto current sm/newmaster.

Why this is a merge with sm/newmaster as the first parent:
yt/Radiation_newmaster only shared an ancient common ancestor with the
mainline (the "Update README.md" commit), so a direct PR to master computed
its three-dot diff from that ancient base and showed ~813 files. In terms of
actual content, yt differs from sm/newmaster in only 192 files -- the genuine
radiation footprint. Recording the integration as a merge whose first parent
is sm/newmaster gives the branch a recent merge-base with master, so a PR from
here shows only the real radiation delta instead of the whole history gap.

Tree content equals yt/Radiation_newmaster (sm/newmaster plus the 192-file
radiation delta): new RT/AMR physics and operators, RadiativeTransfer and
3d_with_rt problem cases, the user_source!/user_primitives! dispatch-signature
refactor (untyped npoin; explicit ::PERT/::TOTAL), and the required RT
integration edits to shared infrastructure (Jexpresso.jl module includes and
eager GridapP4est load, sem_setup cache guards, params_setup atmos_data,
write_output NetCDF, run.jl per-case method clearing). build_rad_2d.jl and
build_rad_3d.jl are consolidated into build_rad.jl.

+++ problems/CompEuler/wave1d_lag/user_source.jl
@@ -1,7 +1,7 @@
 function user_source!(S,
                       q, 
                       qe,
-                      npoin::Int64,
+                      npoin,
                       ::CL, ::TOTAL;
                       neqs=1, x=0.0, y=0.0, xmin=0.0, xmax=1.0)
